+++ azext_iot/common/_azure.py
@@ -7,7 +7,9 @@
 from knack.util import CLIError
 from azext_iot.common.utility import validate_key_value_pairs
 from azext_iot.common.auth import get_aad_token
-from azure.cli.core.commands.client_factory import get_subscription_id
+
+
+IOT_SERVICE_CS_TEMPLATE = "HostName={};SharedAccessKeyName={};SharedAccessKey={}"
 
 
 def _parse_connection_string(cs, validate=None, cstring_type="entity"):
@@ -37,83 +39,6 @@ def parse_iot_device_connection_string(cs):
 def parse_iot_device_module_connection_string(cs):
     validate = ["HostName", "DeviceId", "ModuleId", "SharedAccessKey"]
     return _parse_connection_string(cs, validate, "Module")
-
-
-CONN_STR_TEMPLATE = "HostName={};SharedAccessKeyName={};SharedAccessKey={}"
-
-
-def get_iot_dps_connection_string(
-    cmd,
-    client,
-    dps_name,
-    resource_group_name,
-    policy_name="provisioningserviceowner",
-    key_type="primary",
-):
-    """
-    Function used to build up dictionary of IoT Hub Device
-    Provisioning Service connection string parts
-
-    Args:
-        client ():
-        dps_name (str): Device Provisioning Service name
-        resource_group_name (str): name of Resource Group contianing IoT Hub
-        policy_name (str): Security policy name for shared key; e.g. 'iothubowner'(default)
-        key_type (str): Shared key; either 'primary'(default) or 'secondary'
-
-    Returns:
-        (dict): of connection string elements.
-
-    Raises:
-        CLIError: on input validation failure.
-    """
-    target_dps = None
-    policy = None
-
-    def _find_iot_dps_from_list(all_dps, dps_name):
-        if all_dps:
-            return next(
-                (dps for dps in all_dps if dps_name.lower() == dps.name.lower()), None
-            )
-        return None
-
-    try:
-        target_dps = client.iot_dps_resource.get(dps_name, resource_group_name)
-    except Exception:
-        pass
-
-    if target_dps is None:
-        raise CLIError(
-            "No IoT Provisioning Service found "
-            "with name {} in current subscription.".format(dps_name)
-        )
-
-    try:
-        policy = client.iot_dps_resource.list_keys_for_key_name(
-            dps_name, policy_name, resource_group_name
-        )
-    except Exception:
-        pass
-
-    if policy is None:
-        raise CLIError(
-            "No keys found for policy {} of "
-            "IoT Provisioning Service {}.".format(policy_name, dps_name)
-        )
-
-    result = {}
-    result["cs"] = CONN_STR_TEMPLATE.format(
-        target_dps.properties.service_operations_host_name,
-        policy.key_name,
-        policy.primary_key if key_type == "primary" else policy.secondary_key,
-    )
-    result["entity"] = target_dps.properties.service_operations_host_name
-    result["policy"] = policy_name
-    result["primarykey"] = policy.primary_key
-    result["secondarykey"] = policy.secondary_key
-    result["subscription"] = get_subscription_id(cmd.cli_ctx)
-
-    return result
 
 
 def get_iot_central_tokens(cmd, app_id, token, central_dns_suffix):